requirements.txt
python-telegram-bot==20.7
pygame==2.6.1
python-dotenv
psycopg2-binary
tkcalendar
requests
flask
flask-cors
numpy
scipy
